+++ inc/wpoptimizer/Controller.php
@@ -689,4 +689,14 @@ class Controller
             'error' => 'Missing WPOptimizerController class in theme',
         ];
     }
+
+    public static function maybe_set_default_options(): void {
+        if (false === get_option(self::OPTION_NAME, false)) {
+            $defaults = [];
+            foreach (Settings::get_fields() as $field) {
+                $defaults[$field['id']] = $field['default'];
+            }
+            add_option(self::OPTION_NAME, $defaults);
+        }
+      }
 }